api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.2",
    "langchain-core>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
